requirements.txt
matplotlib
seaborn
pillow
tensorflow
scikit-learn
numpy
opencv-python